Anh Tuan muon nhanh Duy push code

+++ src/main/java/Model/Product.java
@@ -20,7 +20,7 @@ public class Product {
 
   private int quantity;
 
-  private int price;
+  private double price;
 
   private String genre;
 
@@ -63,7 +63,7 @@ public class Product {
       int suppliersId,
       String name,
       int quantity,
-      int price,
+      double price,
       String genre,
       String brand,
       String operatingSystem,
@@ -92,7 +92,7 @@ public class Product {
       int suppliersId,
       String name,
       int quantity,
-      int price,
+      double price,
       String genre,
       String brand,
       String operatingSystem,
@@ -131,7 +131,7 @@ public class Product {
       String.valueOf(id),
       name,
       String.valueOf(quantity),
-      Integer.toString(price),
+      Double.toString(price),
       genre,
       brand,
       operatingSystem,
